requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
iniconfig==2.0.0
Markdown==3.5.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pytest==8.0.2
pytest-django==4.8.0
pytz==2024.1
sqlparse==0.4.4
